backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
Pillow
markdown 
python-multipart